+++ SMLRcam/SMLRcam/FluxPublicProfileViewController.m
@@ -71,7 +71,7 @@
 }
 
 
--(float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath{
+-(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath{
     return 600.0;
 }
